Major Connnfffiiggg
A little adjustment here, and a little adjustment there and oh. Can you change that setting for me?

Config has now been added to change the games play style. Allowing you to change the Highscore limit, the Ambulance speed and health.
Furthermore a sneak peak into the not so far future is also available in the config, with the ability to select maps so you can select which map you want with what narrative your write.

Maps will be added in a future update so keep an eye out !